requirements.txt
# Notev Requirements

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Anthropic Claude API
anthropic==0.39.0
httpx==0.27.0

# Voyage AI (for embeddings)
voyageai==0.2.3

# Document Processing
python-docx==1.1.0
python-pptx==0.6.23
pypdf==3.17.4

# Data Processing
numpy==1.26.2

# Environment Variables
python-dotenv==1.0.0

# WSGI Server (for production)
gunicorn==21.2.0

# File handling
Werkzeug==3.0.1

# Testing (optional, for development)
pytest==7.4.3
pytest-cov==4.1.0
